Logout stay in the same place if pressed cancel

+++ frontend/src/components/Navbar/UserNavBar/dropDownMenu.jsx
@@ -1,24 +1,29 @@
 import React from 'react';
 import styles from './Navbar.css';
+import { Link, redirect } from 'react-router-dom';
+import HomePage from '../../../Page/HomePage/HomePage';
 
 const OwnerDropDownProfile = () => {
-    const changeStatusToNull = () => {
-        window.localStorage.setItem('status', 'null');
-        window.localStorage.removeItem('email');
-    }
+    const changeStatusToNull = (event) => {
+        event.preventDefault();
+        let answer = window.confirm("Are you sure you want to logout?");
+        if (answer) {
+            window.localStorage.setItem('status', 'null');
+            window.localStorage.removeItem('email');
+            window.location.href = '/';
+        }
+    };
+
     return (
         <div className='flex flex-col dropDownProfile'>
             <ul className='flex flex-col gap-4'>
-
-
-                {/* To add an element to the DropDown menu */}
                 <li><a className='Profile' href='/adminprofile'>Profile</a></li>
                 <li><a className='MyProducts' href='/MyProducts'>My Products</a></li>
                 <li><a className='MyCrafts' href='/MyCrafts'>My Crafts</a></li>
-                <li><a className='Logout' href='/' onClick={changeStatusToNull}>Log out</a></li>
+                <li><a className='Logout' onClick={changeStatusToNull}>Log out</a></li>
             </ul>
         </div>
-    )
-}
+    );
+};
 
 export default OwnerDropDownProfile;
